Initial commit: slackrole for the Ganeti Hoogle web site
The slack role installs the Apache web server together with Hoogle and
populates the Hoogle database with the most recent (auto-generated)
Ganeti version, as well as the with the data from the specified
packages Ganeti depends on.

It also sets up a cron job for updating the Ganeti Hoogle database every
hour.

Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>